Multiple Handlers for 1 Promise
We can add multiple .then for one promise and all will work. they will not pass any data to other .then but will work independently

This is different from promise chaining. In promise chaining, .then are bound with each other and also pass on the data to proceed the next.